+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - The last opened file will now be added to the recent documents on start up.
 - The window's title will now only contain the currently opened file's name, and never the full path, even for root files.
 - Add shortcuts for easier access of the recent documents.
+- HTML export now relies on Pandoc, if available, and only falls back to Showdown if Pandoc hasn't been found on the system.
 
 ## Under the Hood